# api/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Rncasp API
  description: >
    REST API for the Rncasp (Really No Clue About Shift Planning) application.
    A shift planning web application for LAN party events that replaces Excel-based
    workflows where team members sign up for shifts across multi-day events.


    All responses (except health and binary downloads) are wrapped in a standard
    envelope: `{"data": ..., "error": ..., "meta": ...}`.


    Authentication is cookie-based using session cookies set on login/register.
  version: 1.0.0
  contact:
    name: Rncasp
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - cookieAuth: []
paths:
...
  /api/teams:
    get:
      tags: [Teams]
      operationId: listTeams
      summary: List all teams
      description: >
        Returns all teams. Non-admin users only see active teams.
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Teams]
      operationId: createTeam
      summary: Create a team
      description: Super-admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [super_admin, user, read_only]
        - name: account_type
          in: query
          schema:
            type: string
            enum: [local, oauth, dummy]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/events/{slug}:
    get:
      tags: [Events]
      operationId: getEvent
      summary: Get an event by slug
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Events]
      operationId: updateEvent
      summary: Update an event
      description: Event admin or super-admin. All fields are optional.
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Events]
      operationId: deleteEvent
      summary: Delete an event
      description: Super-admin only.
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      responses:
        "200":
          description: Event deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/events/{slug}/export/csv:
    get:
      tags: [Export]
      operationId: exportCSV
      summary: Export event shifts as CSV
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{slug}/export/ical:
    get:
      tags: [Export]
      operationId: exportICal
      summary: Export event shifts as iCal
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      responses:
        "200":
          description: iCal file download
          content:
            text/calendar:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{slug}/export/pdf:
    get:
      tags: [Export]
      operationId: exportPDF
      summary: Export event shifts as PDF
      parameters:
        - $ref: "#/components/parameters/EventSlug"
        - name: layout
          in: query
          schema:
            type: string
            enum: [grid, list]
            default: grid
        - name: paper
          in: query
          schema:
            type: string
            enum: [A4, A3]
            default: A4
        - name: landscape
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
        - name: coverage
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
        - name: colors
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
        - name: days
          in: query
          description: Comma-separated ISO dates (YYYY-MM-DD). Empty = all days.
          schema:
            type: string
        - name: users
          in: query
          description: Comma-separated user UUIDs. Empty = all users.
          schema:
            type: string
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Webhooks
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session cookie set by the login or register endpoints.
        The cookie name is configurable; `session` is the default.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    EventSlug:
      name: slug
      in: path
      required: true
      description: URL-friendly event identifier
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 50
        minimum: 1
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        default: 0
        minimum: 0

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: invalid_input
              message: "invalid request body"
    Unauthorized:
      description: Not authenticated or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthorized
              message: "not authenticated"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: forbidden
              message: "insufficient permissions"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: not_found
              message: "resource not found"
    Conflict:
      description: Resource already exists or conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: already_exists
              message: "resource already exists"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - not_found
            - already_exists
            - conflict
            - unauthorized
            - forbidden
            - invalid_input
            - event_locked
            - overbooking_not_allowed
            - account_inactive
            - dummy_account_cannot_login
            - invalid_totp_code
            - totp_required
            - too_many_requests
        message:
          type: string
          description: Human-readable error message
        field:
          type: string
          description: Field name for validation errors
...
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
          required: [message]

    # --- Health ---
...
    User:
      type: object
      required:
        - id
        - username
        - full_name
        - role
        - language
        - account_type
        - time_format
        - totp_enabled
        - is_active
        - created_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        full_name:
          type: string
        display_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        role:
          type: string
          enum: [super_admin, user, read_only]
        language:
          type: string
          enum: [en, de]
        account_type:
          type: string
          enum: [local, oauth, dummy]
        time_format:
          type: string
          enum: ["24h", "12h"]
        totp_enabled:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Team:
      type: object
      required: [id, name, abbreviation, color, sort_order, is_active, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        abbreviation:
          type: string
          description: Single-character team code displayed in the grid
        color:
          type: string
          description: Hex color code (e.g. "#FF5733")
        sort_order:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    TeamResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Team"
    CreateTeamRequest:
      type: object
      required: [name, abbreviation, color]
      properties:
        name:
          type: string
        abbreviation:
          type: string
        color:
          type: string
          description: Hex color code
        sort_order:
          type: integer
          default: 0
...
    Event:
      type: object
      required:
        - id
        - name
        - slug
        - start_time
        - end_time
        - time_granularity
        - is_locked
        - is_public
        - is_event_admin
        - created_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        participant_count:
          type: integer
          nullable: true
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        time_granularity:
          type: string
          enum: ["15min", "30min", "1hour"]
        is_locked:
          type: boolean
        is_public:
          type: boolean
        is_event_admin:
          type: boolean
          description: Whether the requesting user is an admin for this event
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    EventResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Event"
...
    UpdateEventRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        participant_count:
          type: integer
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        time_granularity:
          type: string
          enum: ["15min", "30min", "1hour"]
...